pyproject.toml
[project]
name = "kvrrj"
version = "0.1.0"
description = "Toolkit for doing things ..."
readme = "README.md"
authors = [
    { name = "Tristan Gerrish", email = "Tristan.Gerrish@burohappold.com" }
]
requires-python = ">=3.10"
dependencies = [
    "astral>=3.2",
    "attrs==23.2.0",
    "brotli==1.1.0",
    "case-converter>=1.2.0",
    "click==8.1.7",
    "click-plugins==1.1.1",
    "colorama==0.4.6",
    "contourpy==1.2.1",
    "cycler==0.12.1",
    "dash>=2.18.2",
    "dask[distributed]>=2025.2.0",
    "docutils==0.21.2",
    "dragonfly-core==1.48.7",
    "dragonfly-display==0.2.0",
    "dragonfly-energy==1.27.12",
    "dragonfly-radiance==0.3.116",
    "dragonfly-schema==1.11.4",
    "dragonfly-uwg==0.5.543",
    "flask>=3.0.3",
    "fonttools==4.53.1",
    "geopandas[all]>=1.0.1",
    "geopy>=2.4.1",
    "honeybee-core==1.58.35",
    "honeybee-display==0.4.0",
    "honeybee-doe2==0.22.0",
    "honeybee-energy==1.107.7",
    "honeybee-energy-standards==2.3.0",
    "honeybee-idaice==0.4.14",
    "honeybee-ies==0.12.5",
    "honeybee-radiance==1.66.94",
    "honeybee-radiance-command==1.22.4",
    "honeybee-radiance-folder==2.11.15",
    "honeybee-radiance-postprocess==0.4.420",
    "honeybee-schema==1.58.1",
    "honeybee-standards==2.0.6",
    "inflate64==1.0.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2023.12.1",
    "kiwisolver==1.4.5",
    "ladybug-comfort==0.18.40",
    "ladybug-core==0.43.2",
    "ladybug-display==0.11.2",
    "ladybug-display-schema==0.7.8",
    "ladybug-geometry==1.32.8",
    "ladybug-geometry-polyskel==1.6.7",
    "ladybug-radiance==0.2.10",
    "ladybug-rhino==1.42.15",
    "ladybug-vtk==0.14.3",
    "lbt-dragonfly==0.11.274",
    "lbt-honeybee==0.8.338",
    "lbt-ladybug==0.27.98",
    "lbt-recipes==0.26.9",
    "lockfile==0.12.2",
    "luigi==3.3.0",
    "matplotlib==3.9.1",
    "meteostat>=1.6.8",
    "multivolumefile==0.2.3",
    "numpy==2.0.1",
    "openpyxl>=3.1.5",
    "packaging==24.1",
    "pillow==10.4.0",
    "plotly>=6.0.0",
    "pollination-handlers==0.10.7",
    "psutil==6.0.0",
    "pvlib>=0.12.0",
    "py7zr==0.20.5",
    "pyarrow>=19.0.1",
    "pybcj==1.0.2",
    "pycryptodomex==3.20.0",
    "pydantic==1.10.17",
    "pydantic-openapi-helper==0.2.11",
    "pyephem>=9.99",
    "pyet>=1.3.1",
    "pyparsing==3.1.2",
    "pyppmd==1.0.0",
    "python-daemon==3.0.1",
    "python-dateutil==2.9.0.post0",
    "pyyaml==6.0.1",
    "pyzstd==0.16.0",
    "queenbee==1.28.0",
    "queenbee-local==0.6.3",
    "referencing==0.35.1",
    "rpds-py==0.19.1",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.2",
    "shapely>=2.0.7",
    "six==1.16.0",
    "tables>=3.10.1",
    "tenacity==8.5.0",
    "texttable==1.7.0",
    "timezonefinder>=6.5.9",
    "tornado==6.4.1",
    "tqdm>=4.67.1",
    "typing-extensions==4.12.2",
    "uwg==5.8.13",
    "vtk==9.3.1",
    "xlrd>=2.0.1",
]

[project.scripts]
kvrrj = "kvrrj:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pylint>=3.3.5",
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "sphinx>=8.1.3",
    "sphinx-bootstrap-theme>=0.8.1",
    "sphinxcontrib-fulltoc>=1.2.0",
    "sphinxcontrib-napoleon>=0.7",
]

[tool.pytest.ini_options]
pythonpath = "src"
